Accept BigDecimal in the StripeChargeData (#53)

+++ src/main/scala/model/Currency.scala
@@ -23,7 +23,7 @@ object Currency extends Enum[Currency] with CirceEnum[Currency] {
 
   case object NZD extends Currency
 
-  def exceedsMaxAmount(amount: Int, currency: Currency): Boolean = {
+  def exceedsMaxAmount(amount: BigDecimal, currency: Currency): Boolean = {
     val maxAmount = if (currency.equals(AUD)) 16000 else 2000
     if (amount <= maxAmount) false else true
   }
